backend/requirements.txt
langchain
transformers
torch
faiss-cpu
flask
python-dotenv
sentence-transformers
requests
beautifulsoup4
html5lib
langchain-huggingface